Add extra data in Domain\Transfer\In\Completed event
