+++ lib/event/domain/register/success.php
@@ -19,6 +19,7 @@
 namespace Automattic\Domain_Services_Client\Event\Domain\Register;
 
 use Automattic\Domain_Services_Client\{Entity, Event, Exception, Helper};
+use Automattic\Domain_Services_Client\Response\Event_Aware;
 
 /**
  * Success event for a `Domain\Register` command.
@@ -31,6 +32,10 @@ class Success implements Event\Event_Interface, Event\Async_Command_Related_Inte
 	use Event\Async_Command_Related_Trait;
 	use Event\Object_Type_Domain_Trait;
 	use Event\TransferLocked_Trait;
+	use Event\Renewable_Until_Trait;
+	use Event\Created_Date_Trait;
+	use Event\Expiration_Date_Trait;
+	use Event\Unverified_Contact_Suspension_Date_Trait
 
 	/**
 	 * Gets the date the domain will exit the Add Grace Period (AGP); null if no AGP is offered
@@ -60,50 +65,6 @@ class Success implements Event\Event_Interface, Event\Async_Command_Related_Inte
 			$epp_statuses[] = new Entity\Epp_Status_Code( $epp_status_data );
 		}
 		return new Entity\Epp_Status_Codes( ...$epp_statuses );
-	}
-
-	/**
-	 * Gets the date the domain was created
-	 *
-	 * @return null|\DateTimeInterface
-	 */
-	public function get_created_date(): ?\DateTimeInterface {
-		$created_date = $this->get_data_by_key( 'event_data.created_date' );
-
-		return null === $created_date ? null : Helper\Date_Time::createImmutable( $created_date );
-	}
-
-	/**
-	 * Gets the domain expiration date
-	 *
-	 * @return null|\DateTimeInterface
-	 */
-	public function get_expiration_date(): ?\DateTimeInterface {
-		$expiration_date = $this->get_data_by_key( 'event_data.expiration_date' );
-
-		return null === $expiration_date ? null : Helper\Date_Time::createImmutable( $expiration_date );
-	}
-
-	/**
-	 * Get the last date to renew the domain
-	 *
-	 * @return null|\DateTimeInterface
-	 */
-	public function get_renewable_until(): ?\DateTimeInterface {
-		$renewable_until = $this->get_data_by_key( 'event_data.renewable_until' );
-
-		return null === $renewable_until ? null : Helper\Date_Time::createImmutable( $renewable_until );
-	}
-
-	/**
-	 * Gets the date when the domain will be suspended if the contact information is not verified
-	 *
-	 * @return null|\DateTimeInterface
-	 */
-	public function get_unverified_contact_suspension_date(): ?\DateTimeInterface {
-		$unverified_contact_suspension_date = $this->get_data_by_key( 'event_data.unverified_contact_suspension_date' );
-
-		return null === $unverified_contact_suspension_date ? null : Helper\Date_Time::createImmutable( $unverified_contact_suspension_date );
 	}
 
 	/**
