+++ lib/event/expiration-date-trait.php
@@ -16,22 +16,22 @@
  * if not, see https://www.gnu.org/licenses.
  */
 
-namespace Automattic\Domain_Services_Client\Event\Domain\Transfer\In;
+namespace Automattic\Domain_Services_Client\Event;
 
-use Automattic\Domain_Services_Client\{Event};
+use Automattic\Domain_Services_Client\{Helper};
 
 /**
- * Inbound domain transfer success event
- *
- * This event is generated when a domain transfer from another registrar to the reseller's account is successful.
+ * Trait that adds the `get_expiration_date` method to an event
  */
-class Completed implements Event\Event_Interface, Event\Async_Command_Related_Interface {
-	use Event\Async_Command_Related_Trait;
-	use Event\Object_Type_Domain_Trait;
-	use Event\Transfer_Trait;
-	use Event\TransferLocked_Trait;
-	use Event\Renewable_Until_Trait;
-	use Event\Created_Date_Trait;
-	use Event\Expiration_Date_Trait;
-	use Event\Unverified_Contact_Suspension_Date_Trait
+trait Expiration_Date_Trait {
+	/**
+	 * Gets the domain expiration date
+	 *
+	 * @return null|\DateTimeInterface
+	 */
+	public function get_expiration_date(): ?\DateTimeInterface {
+		$expiration_date = $this->get_data_by_key( 'event_data.expiration_date' );
+
+		return null === $expiration_date ? null : Helper\Date_Time::createImmutable( $expiration_date );
+	}
 }
